move traits to `gears`, update modules
