+++ x/staking/src/abci_handler.rs
@@ -1,8 +1,7 @@
 use crate::{
-    BankKeeper, GenesisState, Keeper, KeeperHooks, Message, QueryDelegationRequest,
-    QueryDelegationResponse, QueryParamsResponse, QueryRedelegationRequest,
-    QueryRedelegationResponse, QueryUnbondingDelegationResponse, QueryValidatorRequest,
-    QueryValidatorResponse,
+    GenesisState, Keeper, Message, QueryDelegationRequest, QueryDelegationResponse,
+    QueryParamsResponse, QueryRedelegationRequest, QueryRedelegationResponse,
+    QueryUnbondingDelegationResponse, QueryValidatorRequest, QueryValidatorResponse,
 };
 use gears::{
     context::{block::BlockContext, init::InitContext, query::QueryContext, tx::TxContext},
@@ -16,7 +15,10 @@ use gears::{
             begin_block::RequestBeginBlock, end_block::RequestEndBlock, query::RequestQuery,
         },
     },
-    x::{keepers::auth::AuthKeeper, module::Module},
+    x::{
+        keepers::{auth::AuthKeeper, bank::StakingBankKeeper, staking::KeeperHooks},
+        module::Module,
+    },
 };
 use serde::Serialize;
 
@@ -25,7 +27,7 @@ pub struct ABCIHandler<
     SK: StoreKey,
     PSK: ParamsSubspaceKey,
     AK: AuthKeeper<SK, M>,
-    BK: BankKeeper<SK, M>,
+    BK: StakingBankKeeper<SK, M>,
     KH: KeeperHooks<SK, AK, M>,
     M: Module,
 > {
@@ -56,7 +58,7 @@ impl<
         SK: StoreKey,
         PSK: ParamsSubspaceKey,
         AK: AuthKeeper<SK, M>,
-        BK: BankKeeper<SK, M>,
+        BK: StakingBankKeeper<SK, M>,
         KH: KeeperHooks<SK, AK, M>,
         M: Module,
     > ABCIHandler<SK, PSK, AK, BK, KH, M>
