--- x/staking/src/external/mod.rs
@@ -1,1 +1,0 @@
-mod bank;
